feat(worldgen): add layout_core.js with buildBaseTown and wire into town_gen.js
